Rearrange some code into modules
